requirements.txt
fastapi
uvicorn
ollama
faiss-cpu
sentence-transformers
requests
